Fixed compiler warnings in bulk1.c and features-10_2.c

+++ unittest/libmariadb/features-10_2.c
@@ -86,7 +86,7 @@ static int execute_direct_example(MYSQL *mysql)
   bind[0].buffer= &intval;
   bind[1].buffer_type= MYSQL_TYPE_STRING;
   bind[1].buffer= (char *)strval;
-  bind[1].buffer_length= strlen(strval);
+  bind[1].buffer_length= (unsigned long)strlen(strval);
 
   /* set number of parameters */
   rc= mysql_stmt_attr_set(stmt, STMT_ATTR_PREBIND_PARAMS, &param_count);
